less warnings

+++ src/ast/ast_allocate_stack.cpp
@@ -73,7 +73,7 @@ namespace das {
                 blocks.push_back(block);
             }
             if ( block->arguments.size() || block->copyOnReturn || block->moveOnReturn ) {
-                auto sz = sizeof(BlockArguments);
+                auto sz = uint32_t(sizeof(BlockArguments));
                 block->stackTop = allocateStack(sz);
                 if ( log ) {
                     logs << "\t" << block->stackTop << "\t" << sz
